% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tract_from_route.R
\name{census_tract_from_route}
\alias{census_tract_from_route}
\title{Get all the census tracts that are served by a King Country Metro route}
\usage{
census_tract_from_route(route)
}
\arguments{
\item{route}{A string that specifies a King County Metro route.}
}
\value{
A character vector that represents all the US Census Bureau census tracts that are
served by that route.
}
\description{
Takes a string that specifies a King Country Metro route and returns a
character vector of all the US Census Bureau census tracts that are
served by that route.
}
\examples{

census_tract_from_route( 45 ) 
census_tract_from_route( "D Line" ) 
census_tract_from_route( "LINK" ) 

}

