% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tract_from_stop.R
\name{census_tract_from_stop}
\alias{census_tract_from_stop}
\title{Get the census tract that contains a King Country Metro stop}
\usage{
census_tract_from_stop(stop)
}
\arguments{
\item{stop}{A string that specifies a King County Metro stop.}
}
\value{
Returns a string that represents that ID for the US Census Bureau census tract
that contains the specified stop.
}
\description{
Takes a string that specifies a King Country Metro route and returns a string that 
represents the ID for the US Census Bureau census tract that contains
that stop.
}
\examples{

#
# 8th Ave NW & NW 80th Street
#
census_tract_from_stop( 28100 ),

#
# Denny & Aurora
#
census_tract_from_stop( 6235 ),

}

