% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kct.shp.proj.R
\docType{data}
\name{kct.shp.pro}
\alias{kct.shp.pro}
\title{Geospatial data for US Census Bureau census tracts for King County,
Washington.}
\format{}
\source{
See the following URL. Near the bottom of the page is a section titled
*ArcGIS Shapefiles* which has links to download the data files.

  <http://www.seattle.gov/dpd/cityplanning/populationdemographics/geographicfilesmaps/2010census/default.htm>
}
\usage{
data( kct.shp.pro )
}
\description{
Geospatial data for US Census Bureau census tracts for King County,
Washington.
}
\keyword{datasets}

