% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tract.R
\name{plot_tract}
\alias{plot_tract}
\title{Plot all the transit-stop locations for a census tract on a map}
\usage{
plot_tract(tract, map.zoom = 14)
}
\arguments{
\item{tract}{A string that specifies a US Census Bureau census tract.}

\item{map.zoom}{Zoom level for the map. Larger numbers are closer to
the earth. Default is 14. Maximum value is 21.}
}
\value{
None. Plot a map that shows the specified census tract and
all the King County Metro transit stops in that tract.
}
\description{
Takes a census tract, plots it on a map, and draws all King County
Metro transit stops that are contained within that tract.
}
\examples{

plot( 30, map.zoom = 15 )
plot( 53.01 )


Get the lon/lat of the central point of the census tract.

}

