% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.stops.df.R
\docType{data}
\name{route.stops.df}
\alias{route.stops.df}
\title{Mapping of King County Metro routes to stops}
\format{The dataset is two columns, `routes` and `stops`. The
relationship it many-to-many: a route serves multiple stops and a given
route might served by multiple routes.}
\source{
This data was extracted from the transit description files
published by King Country. 

You can find these transit description files that describe route and stops for King County
Metro bus transit at the following URL.

  <http://metro.kingcounty.gov/GTFS/>

These files are in a format developed by Google for describing transit
systems.  You can find instructions for how to interpret the 
data files at the following URL:

  <https://developers.google.com/transit/gtfs/reference?csw=1>
}
\usage{
data( route.stops.df )
}
\description{
Mapping of King County Metro routes to stops
}
\keyword{datasets}

