% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.demographics.kc.R
\docType{data}
\name{tract.demographics.kc.R}
\alias{tract.demographics.kc.R}
\title{Demographic data related to population and housing from the
Washington State Office of Financial Management.}
\format{The data set has the following columns.

- County.Name
- Census.Tract
- Total.Population
- Population.Density.Persons.Square.Mile
- Total.Housing.Units
- Occupied.Housing.Units
- Vacant.Housing.Units
- Occupancy.Rate
- Vacancy.Rate
- Land.Area.Square.Miles
- Total.Area.Square.Miles
- Water.Area
- Geoid

Of these columns, however, only the following have been converted 
into numeric values that are usable.

- Census.Tract
- Total.Population
- Population.Density.Persons.Square.Mile
- Total.Housing.Units
- Land.Area.Square.Miles}
\source{
This data was extracted from the follow Excel file

    wa_2010_pl94171_tract.xlsx

which was downloaded from the the `Census tract` link at the following
web page.

\url{http://www.ofm.wa.gov/pop/census2010/data.asp}
}
\usage{
data( tract.demographics.kc.R )
}
\description{
Demographic data related to population and housing from the
Washington State Office of Financial Management.
}
\keyword{datasets}

